requirements.txt
# ------------------------
# Core App
# ------------------------
streamlit
python-dotenv

# ------------------------
# LangChain Core
# ------------------------
langchain
langchain-core
langchain-community
langchain-text-splitters

# ------------------------
# LLMs
# ------------------------
google-generativeai
langchain-google-genai
langchain-ollama

# ------------------------
# Embeddings & Vector DB
# ------------------------
sentence-transformers
huggingface-hub
faiss-cpu

# ------------------------
# PDF & OCR
# ------------------------
PyPDF2
pytesseract
pdf2image
Pillow
